Fix download AGCOM list and improvements
We fixed the download of the AGCOM list. Now we download just the latest "Determina" from their website.

We centralized the blacklists IPs into the censorship_params.sh file and the path for the manual list.

We also introduced a check if the manual list file exists so that it doesn't throw an exception when it doesn't.

Added the processing of the AGCOM list in the update_blacklists.sh file.